pyproject.toml
[tool.poetry]
name = "companalysis"
version = "0.1.0"
description = ""
authors = ["Saber Zerhoudi <saber.zerhoudi@uni-passau.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tqdm = "^4.66.4"
pandas = "^2.2.2"
matplotlib = "^3.8.4"
ijson = "^3.2.3"
nltk = "^3.8.1"
seaborn = "^0.13.2"
openai = "^1.30.1"
argparse = "^1.4.0"
python-dotenv = "^1.0.1"
migrate = "^0.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
